Access the drones patent dataset

A set of 18,970 patent publications containing the word drone or drones published between 1845 and the 2nd of November 2017 from the Clarivate Analytics Derwent Innovation database.

data("drones")

Format

A data frame with 18,970 observations of 22 variables:

abstract

The original document abstract, 12798 93 percent of documents, a character vector

abstract_english

The english document abstract, 12295 94 percent of documents, a character vector

application_number

The long application number including the date, 15776 100 percent of documents, a character vector

applicant

The patent applicant name, also known as the assignee name, 7746 89 percent of documents, a character vector

applicant_cleaned

A cleaned version of the applicant name, 6,929 names, a character vector

cited_nonpatent

Literature citations, field is noisy, 29599, 36 percent of documents, a character vector

cited_patents

Patents cited in one or more documents, 93769 39 percent of documents, a character vector

citing_patents

Patents citing one or more documents, 79406 39 percent of documents, a character vector

cpc

The Cooperative Patent Classification Codes, 17077 93 percent of documents, a character vector

dwpi_family_dates

Family dates for DWPI family numbers, 5154 94 percent of documents, a character vector

dwpi_family_kind

Document kind codes for DWPI Family members, 42 94 percent of documents, a character vector

dwpi_family_numbers

DWPI family members - Derwent World Patent Index -, 31007 94 percent of documents, a character vector

first_claim

The first claim in a patent document, 15376 97 percent of documents, a character vector

inpadoc_family_members

INPADOC Family Members in long format with dates, 49625 97 percent of documents, a character vector

inpadoc_first_family_member

The earliest publication number in the inpadoc_family_members based on the date, 9,020, 98 percent of documents, a character vector

inventor

The original inventor name, 19630 94 percent of documents, a character vector

ipc

International Patent Classification - IPC - codes, 8727 98 percent of documents, a character vector

priority_number

Patent priority numbers in long format with dates, 23382 99 percent of documents, a character vector

publication_number

Publication numbers in short form minus dates, 18970 100 percent of documents, a character vector

publication_year

The year of publication of the publication numbers, 145 99 percent of documents, a character vector

related_application_numbers

Details of related patent applications, 7906 38 percent of documents, a character vector

title_english

The english title, 11113 99 percent of documents, a character vector

title_original

The original title, normally concatenated as English, French, German etc, 14324 97 percent of documents, a character vector

Source

Clarivate Analytics Derwent Innovation database

Details

Patent data fields are commonly concatenated with a semicolon and require tidying for accurate counts. The cited_nonpatent field in this dataset contains irrelevant legal status information and is messy. Applicant names (assignees) were cleaned using VantagePoint by fuzzy matching names grouped on the priority number followed by manual review. In the second step the cleaned data was fuzzy match grouped on the INPADOC family member number.