Access drones patent data from the Lens

A set of 9071 patent documents from the open access Lens patent database containing the word drone or drones in the full text of documents published between 1948 and the 31st of December 2017 and reduced from 16,419 publications to one record per family. The records can be accessed online at the public collection https://www.lens.org/lens/collection/45155.

data("lens_drones")

Format

A data frame with 9071 observations of 24 variables:

number

The record number (row no), a character vector

jurisdiction

The two letter country or instrument code where a document was published, a character vector

kind

The kind code describing the publication level of a document, a character vector

publication_number

The publication number including forward /, a character vector

lens_id

The internal Lens database identifier, a character vector

publication_date

a character vector

publication_year

a character vector

application_number

a character vector

application_date

a character vector

priority_numbers

Patent priority numbers in long format with dates, a character vector

title

The patent document title, a character vector

applicants

patent applicants concatenated with ";;", a character vector

inventors

The original inventor name, concatenated with ";;", a character vector

url

url link to the record, a character vector

type

A Patent Application, Granted Patent or other type, a character vector

full_text

Indicates whether the full text is available at the Lens, a character vector

cited_count

Number of later patent filings citing the document, a character vector

simple family size

Assumed to be the DOCDB family based on shared priority, a character vector

extented_family_size

Assumed to be the extended INPADOC family but not documented by the Lens, a character vector

sequence_count

A count of DNA or amino acid sequences, a character vector

cpc_classification

The Cooperative Patent Classification Codes, a character vector

ipcr_classifications

International Patent Classification - IPC - codes, a character vector

npl_citation_ids

Non patent literature citation identifiers, a character vector

Source

The Lens Patent Databse

Details

Field names in this dataset have been simplified from their original long form to make them easier to work with in R. Lens Patent data fields are concatenated with a double semicolon and require tidying for accurate counts. Applicant and inventor names have not been cleaned.