patent numbers

A set of 15,776 patent application numbers and associated patent numbers for use in generating statistics with the drones dataset from the Clarivate Analytics Derwent Innovation.

data("numbers")

Format

A data frame with 15,776 observations of 15 variables:

application_number

The long application number including the date, a character vector

basic_patent_number

DWPI basic patent number forming the base of the dwpi_family_numbers, a character vector

basic_patent_date

DWPI basic patent date, a character vector

cited_patents

Patents cited in one or more documents, a character vector

citing_patents

Patents citing one or more documents, a character vector

dwpi_family_dates

Family dates for DWPI family numbers, a character vector

dwpi_family_kind

Document kind codes for DWPI Family members,, a character vector

dwpi_family_numbers

DWPI family members - Derwent World Patent Index -, a character vector

dwpi_family_members_count

DWPI family members count- Derwent World Patent Index -, a character vector

inpadoc_family_members

INPADOC Family Members in long format with dates, a character vector

inpadoc_first_family_member

The earliest publication number in the inpadoc_family_members based on the date, a character vector

priority_number

Patent priority numbers in long format with dates, a character vector

publication_number

Publication numbers in short form minus dates, a character vector

publication_year

The year of publication of the publication numbers, a character vector

related_application_numbers

Details of related patent applications, a character vector

Source

Clarivate Analytics Derwent Innovation

Details

Patent data fields are commonly concatenated with a semicolon and require tidying for accurate counts.